Add a kind annotation on Graph'
  data Graph' block (n :: * -> * -> *) e x where
    GNil  :: Graph' block n O O
    GUnit :: block n O O -> Graph' block n O O
    GMany :: MaybeO e (block n O C)

We want the kind annotation so we can get the the right kind for
Graph' without looking at its *uses*, as Haskell dictates (Report
section 4.6), and (now that we are extending the kind system) GHC is
going to do.

Previously GHC was a bit more forgiving, in a non-standard way.