Migrate testing/Main.hs to TestFramework
This is (hopefully) a first step to cleaning up and extending
Hoopl's testsuite. In the future I'd like to add some QuickCheck
tests, so I wanted to use a testing framework that would make
it easy to handle both HUnit and QuickCheck tests.

Note that we considered using tasty (instead of test-framework),
but it doesn't support GHC <7.4, which Hoopl does support.